+++ examples/custom_vpc/main.tf
@@ -37,5 +37,6 @@ module "instance" {
 
   user_data            = module.user_data.install_sh
   iam_instance_profile = module.agentless_scanner_role.profile.name
+  vpc_id               = var.vpc_id
   subnet_id            = var.subnet_id
 }