+++ modules/instance/variables.tf
@@ -28,6 +28,11 @@ variable "user_data" {
   default     = null
 }
 
+variable "vpc_id" {
+  description = "The VPC ID to launch in"
+  type        = string
+}
+
 variable "subnet_id" {
   description = "The VPC Subnet ID to launch in"
   type        = string